fixing the gap issue on the table view
On iOS 6 on iPads using the frame instead of the bounds added a 20
pixel gap between the top of the tableview and the bottom of the
navigation bar.  I tested this fix on iOS 7 and iOS 6 with both iPads
and iPhones and it seems to have no negative side effects.